Add images to about and menu tabs
