+++ src/scripts/index.js
@@ -1,25 +1,10 @@
 import '../scss/styles.scss';
+import tab from './modules/tab';
+// import navbar from './modules/navbar';
+// import tabbedContainer from './modules/tabbed-container';
+// import footer from './modules/footer';
 
-
-// TABS
-
-function openPage(id) {
-  document.querySelectorAll('.menu__tab').forEach(tab => {
-    tab.classList.remove('active')
-  })
-  document.querySelectorAll('.menu__item').forEach(item => {
-    item.classList.remove('active')
-  })
-
-  document.getElementById(id).classList.add('active')
-  document.querySelector(`div[data-id='${ id }']`).classList.add('active')
-}
-
-document.querySelectorAll('.menu__tab').forEach(tab => {
-  tab.addEventListener('click', e => {
-    openPage(e.target.getAttribute('data-id'))
-  })
-})
-
-
-document.getElementById("default-open").click();
+// tab();
+// navbar();
+// tabbedContainer();
+// footer();
